Split Policy<TResult> classes into separate source files
